requirements.txt
openai
jsonref
requests
pyyaml
AzureOpenAI
jsonpickle